chore: adapt old queries to new models and rename getters to fetchers
