--- front/app/api/selected_communities/types.ts
@@ -1,7 +1,0 @@
-import { CommunityType } from '@/utils/types';
-
-export type CommunitiesParamsOptions = {
-  type: CommunityType | undefined;
-  limit: number | undefined;
-  siren: `${string}` | undefined;
-};
